Create particles.py
This module contains two classes, PointParticle and ParticleSet.

A PointParticle instance is a point particle (particle with
negligible mass). Each point particle has an electric field,
which can be calculated with electric_field method.

A ParticleSet instance is a set of PointParticle instances.
A particle set has an electric field which depends on it's
point particles. This electric field can be calculated with
electric_field method.